+++ contracts/interfaces/ICVIOracle.sol
@@ -2,8 +2,8 @@
 pragma solidity ^0.8;
 
 interface ICVIOracle {
-    function getCVIRoundData(uint80 roundId) external view returns (uint16 cviValue, uint256 cviTimestamp);
-    function getCVILatestRoundData() external view returns (uint16 cviValue, uint80 cviRoundId, uint256 cviTimestamp);
+    function getCVIRoundData(uint80 roundId) external view returns (uint32 cviValue, uint256 cviTimestamp);
+    function getCVILatestRoundData() external view returns (uint32 cviValue, uint80 cviRoundId, uint256 cviTimestamp);
 
     function setDeviationCheck(bool newDeviationCheck) external;
     function setMaxDeviation(uint16 newMaxDeviation) external;
